diff-highlight: use hunk lengths to prevent overrun
The current handle_line function doesn't respect patch hunk lengths.
This means that when there is a commit that changes the last line of
a file, diff-highlight will incorrectly re-order the signature in the
output of `git format-patch`.

For example:
...
-# Last lnie
+# Last line
--
2.19.2

will currently be formatted as:
...
-# Last lnie
--
+# Last line
2.19.2

Reading the hunk lengths and only processing the lines that are
part of the patch preserves the location of the signature.

This also allows the output to be flushed at the end of a hunk,
allowing the user to see the patch while waiting for git to output
the next commit.

Signed-off-by: Norman Rasmussen <norman@rasmussen.co.za>